requirements.txt
alembic==1.8.1
attrs==22.1.0
click==8.1.3
coverage==6.5.0
exceptiongroup==1.0.4
Flask==2.2.2
Flask-HTTPAuth==4.7.0
greenlet==2.0.1
importlib-metadata==5.0.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
marshmallow==3.19.0
mysql-connector-python==8.0.31
mysqlclient==2.1.1
packaging==21.3
pluggy==1.0.0
protobuf==3.20.1
pyparsing==3.0.9
pytest==7.2.0
SQLAlchemy==1.4.44
tomli==2.0.1
Werkzeug==2.2.2
zipp==3.10.0
